seperate file for augmented reader
